// src/app/shared/models/org.model.ts
export class Org {
    id: number;
    first_name: string;
    last_name: string;
    username: string;
    org_name: string;
    contact_name: string;
    email: string;
    phone: number;
    mobile_phone: number;
    fax: number;
    contact_method: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    county: string;
    zip: number;
    password: string;
    isOrg: number;
}